Move TimerServiceFactory to package "service"
