+++ src/main/java/org/example/service/TimerServiceFactory.java
@@ -1,8 +1,8 @@
-package org.example.model;
+package org.example.service;
 
+import org.example.model.AppContext;
 import org.example.service.interfaces.TimerService;
-import org.example.service.InMemoryTimerService;
-import org.example.service.RemoteTimerService;
+
 // this class handles which timerservice class is going to be used
 public final class TimerServiceFactory {
     private static final InMemoryTimerService GUEST = new InMemoryTimerService();
